+++ app/BuildException.php
@@ -1,0 +1,8 @@
+<?php
+namespace App;
+
+class BuildException extends \Exception
+{
+
+	public array $errors;
+}